# back/docs/swagger.yaml
definitions:
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      password:
        type: string
      roles:
        items:
          type: string
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
  utils.HttpError:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
paths:
  /users/:
    get:
      consumes:
      - application/json
      description: get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
      summary: get all users
      tags:
      - user
    post:
      consumes:
      - application/json
      description: create user
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
      summary: create user
      tags:
      - user
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: delete user
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
      summary: delete user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: get users by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HttpError'
      summary: get users by id
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: update user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HttpError'
      summary: update user
      tags:
      - user
swagger: "2.0"